# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Practica de laravel",
        "version": "2.0"
    },
    "paths": {
        "/api/school": {
            "get": {
                "tags": [
                    "Colegio"
                ],
                "summary": "Obtener lista de colegios",
                "operationId": "e608966a75b1a16f53270325f3170292",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una lista de profesores"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Colegio"
                ],
                "summary": "crear una escuela",
                "operationId": "cabeadfa95539261d8c761d103c8fb43",
                "requestBody": {
                    "description": "Datos necesarios para crear una escuela",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gregorio luperon"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Escuela publica"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crea un estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/school/{id}": {
            "get": {
                "tags": [
                    "Colegio"
                ],
                "summary": "Obtener un escuelas",
                "operationId": "96b0fe3b79f8a2b8edf0c02608f9693b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Colegio"
                ],
                "summary": "actualizar una escuela",
                "operationId": "b7c3412840c1344c2d122700074990e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para actualizar un estudiante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Una estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Colegio"
                ],
                "summary": "eliminar un escuela",
                "operationId": "dff80df27ee18470c168d00d8fcd4845",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una Escuela"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
...
        "/api/teacher": {
            "get": {
                "tags": [
                    "Profesor"
                ],
                "summary": "Obtener lista de profesores",
                "operationId": "11724500c723f6203a0dd605a6e458a8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una lista de profesores"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profesor"
                ],
                "summary": "crear un profesor",
                "operationId": "e10b12f0f4279b43cf381d4f14cdc249",
                "requestBody": {
                    "description": "Datos necesarios para crear un profesor",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gregorio luperon"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Bonaire Encarnacion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crea un profesor"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/teacher/{id}": {
            "get": {
                "tags": [
                    "Profesor"
                ],
                "summary": "Obtener un profesor",
                "operationId": "c0859286cf622737ae5472567c20b138",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un profesor"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profesor"
                ],
                "summary": "actualizar una profesor",
                "operationId": "bc0a4dc01efdf593e402df77c96e5ca1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para actualizar un profesor",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jose"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Encarnacion"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un profesor"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profesor"
                ],
                "summary": "eliminar un profesor",
                "operationId": "eba7abb322134e0a4d89d2a6b2caa641",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un Profesor"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        }
    },
...
}